Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 129ms, Memory: 10MB, Status: Passed - Codetree

+++ 250603/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -27,9 +27,9 @@ public class Main {
     public static void main(String[] args) throws IOException {
         Scanner sc = new Scanner(System.in);
 
-        m = sc.nextInt();
-        long a = sc.nextInt();
-        long b = sc.nextInt();
+        m = sc.nextLong();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         long mmax = Long.MIN_VALUE;
         long minx = Long.MAX_VALUE;
         for (long i = a; i <= b; i++) {
